// src/app/models/usuario.ts
export class Usuario {
    id?: string;
    login?: string;
    identificacion?: number;
    foto?:string;
    password?: string;
    telefono?: string;
    perfil?: string;
    estado?: string;
    nombre?: string;
    direccion?: string;
    apellidos?: string;
    email?: string;
    fecha_create?: number;
    fecha_update?: number;
    usuario_update?: string;
    puntos?:number;
    categoria?:string;
    roles?:any[];
}